V1.1 fixed mv

+++ MAKEFILE.sh
@@ -16,17 +16,27 @@ apt install htop ripgrep fzf neofetch git gh zoxide curl unzip wget tldr -y
 curl -s https://ohmyposh.dev/install.sh | bash -s
 
 mkdir -p $DIR/.dotfiles/archive/
+if [ -f $DIR/.bashrc ]; then
+  mv $DIR/.bashrc $DIR/.dotfiles/archive/.bashrc
+  ln -sf $DIR/.dotfiles/.bashrc $DIR
+fi
 
-mv $DIR/.bashrc $DIR/.dotfiles/archive/.bashrc
-ln -sf $DIR/.dotfiles/.bashrc $DIR
+if [ -f $DIR/.profile ]; then
+  mv $DIR/.profile $DIR/.dotfiles/archive/.profile
+  ln -sf $DIR/.dotfiles/.profile $DIR
+fi
 
-mv $DIR/.profile $DIR/.dotfiles/archive/.profile
-ln -sf $DIR/.dotfiles/.profile $DIR
+if [ -f $DIR/.config/neofetch/conf.conf ]; then
+  mkdir -p $DIR/.dotfiles/archive/neofetch
+  mkdir -p $DIR/.config/neofetch
+  mv $DIR/.config/neofetch/config.conf $DIR/.dotfiles/archive/neofetch/config.conf
+  ln -sf $DIR/.dotfiles/neofetch/config.conf $DIR/.config/neofetch
+fi
 
-mkdir -p $DIR/.dotfiles/archive/neofetch
-mv $DIR/.config/neofetch/config.conf $DIR/.dotfiles/archive/neofetch/config.conf
-ln -sf $DIR/.dotfiles/neofetch/config.conf $DIR/.config/neofetch
+if [ -f $DIR/.gitconfig ]; then
+  mkdir -p $DIR/.dotfiles/archive/git
+  mv $DIR/.gitconfig $DIR/.dotfiles/archive/git/.gitconfig
+  ln -sf $DIR/.dotfiles/git/.gitconfig $DIR
+fi
 
-mkdir -p $DIR/.dotfiles/archive/git
-mv $DIR/.gitconfig $DIR/.dotfiles/archive/git/.gitconfig
-ln -sf $DIR/.dotfiles/git/.gitconfig $DIR
+exec bash
